Kharazim added, popularity/win percentage bug fix
Added Kharazim to the list of heroes and thus a wiki is created with his
top builds. The highest popularity and highest win percentage talents were
mistakenly swapped, now they should be in the correct place.

Performed a number of refactorings in order to make the code PEP 8
compliant. Further changes were necessary to run the project under Python
3. Cosmetic changes were needed in some instances to reinforce clarity.